Add IS452 Midterm Project
This is my midterm project for IS452. It's just some code that I'm proud because I spent a decent amount of time on it. The goal of the project was to break each chapter of Dracula into its own text file in two different ways. One way just listing Dracula and the chapter number and the other way was listing that as well as the chapter name.